Выполнил задание 'Модуляция' из 5 раздела
