+++ js/filter-form.js
@@ -1,0 +1,1 @@
+//Здесь будет производится фильтрация расположенных рядом объявлений.Как зависимость сюда будет подключаться server.js
