+++ js/map.js
@@ -1,0 +1,1 @@
+//Как зависимость сюда подключается data.js (с полученными с сервера данными). На основании этих данных отрисовываем пины на карте и информацию об этих объявлениях.
