+++ js/popup.js
@@ -1,0 +1,1 @@
+//Здесь будет происходить показ попапа при успешной отправке данных на сервер либо ошибке при отправке. Как зависимость сюда будет подключаться server.js
