+++ js/server.js
@@ -1,0 +1,1 @@
+//Тут пока самый сложный момент. Здесь так понимаю должно быть описано взаимодействие: 1. активация карты - первоначальная загрузка похожих объявлений рядом (data.js возможно надо будет убрать как переходное звено к map.js); 2. при использовании фильтра показываем новые объявления на карте; 3. при отправке объявления пользователем показываем нужный попап
