+++ js/user-form.js
@@ -1,0 +1,1 @@
+//Здесь будет происходить валидация формы заполнения объявления на стороне клиента. Как зависимость сюда будет подключаться server.js
